Patched Datastream#stream to work with controlGroup E (fixes #88)
The `stream` method relied on the `dsSize` attribute of the datastream to report
the content length; however, a Fedora 3 repo always reports `dsSize` 0 for
external datastreams. This patch permist the existing method to work with
external datastreams while preserving the original behavior for other type
(managed and inline).  Since Hydra::Controller::DownloadBehavior relies on this
method, the patch allows that module to function on external datastreams without
modification.