Add Dockerfile and install.properties for fcrepo3
This may not be the ultimate home for these files, but included are a
simple Dockerfile and install.properties for noninteractive installation
of fcrepo3. Running the container starts Tomcat on port 8983, as is
generally expected in the test suite.

There are also a couple of basic Rake tasks to help with building and
running against the container.